requirements.txt
--index-url https://pypi.ccnmtl.columbia.edu/
Django==1.9.7
feedparser==5.2.1
Markdown==2.6.6
simplejson==3.8.2
smartypants==1.8.6
uuid==1.30
psycopg2==2.6.1
Pillow==3.3.0
versiontools==1.9.1
statsd==3.2.1
pep8==1.7.0
pyflakes==1.2.3
mccabe==0.5.0
pycodestyle==2.0.0
flake8==2.6.2
tornado==4.3
BeautifulSoup==3.2.1
cssselect==0.9.2
lxml==3.6.0
fuzzywuzzy==0.11.0
sure==1.4.0
ipython==4.2.0
ipdb==0.10.1
lettuce==0.2.19ccnmtl
rdflib==4.2.1
selenium==2.48.0
coverage==4.1
pyasn1==0.1.9
pyOpenSSL==16.0.0
ndg-httpsclient==0.4.1
requests==2.10.0
logilab-common==1.2.2
logilab-astng==0.24.3
pylint==1.5.6
six==1.10.0
ipaddress==1.0.16
fake-factory==0.5.8
factory_boy==2.7.0
ldap3==1.3.1
contextlib2==0.5.3
rcssmin==1.0.6
rjsmin==1.0.12
Jinja2==2.8
Markupsafe==0.23

django-bootstrap3==7.0.1
djangowind==0.15.0
sorl==3.1
typogrify==2.0.7
django-staticmedia==0.2.2
django-indexer==0.3.0
django-templatetag-sugar==1.0
django-paging==0.2.5
django-appconf==1.0.2
django-compressor==2.0
django-statsd-mozilla==0.3.16
raven==5.21.0
django-bootstrap-form==3.2.1
sqlparse==0.1.19
django-debug-toolbar==1.4
django-waffle==0.11.1
django-jenkins==0.19.0
django-smoketest==1.0.0
django-extensions==1.6.7
django-stagingcontext==0.1.0
django-ga-context==0.1.0
django-impersonate==1.0.1
django-treebeard==4.0.1
django-pagetree==1.2.1
django-pageblocks==1.0.4
django-quizblock==1.2.1
django-markwhat==1.5
djangorestframework==3.2.4
gunicorn==19.6.0
boto==2.41.0
python-dateutil==2.5.3
django-storages-redux==1.3.2
django-cacheds3storage==0.1.2

ccnmtlsettings==0.3.0
nameparser==0.4.0
httplib2==0.9.2
oauth==1.0.1
oauth2==1.9rc1
oauthlib==0.6.3
pylti>=0.1.3
django-lti-provider==0.1.0
